Deepfake Detection Tool/requirements.txt
streamlit
python-dotenv
openai
requests
pillow
moviepy>=2.0
imageio-ffmpeg